code-analysis-service/src/middleware/auth.py
"""
🔒 Authentication Middleware - JWT verification for protected routes
"""

from fastapi import Request, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
import jwt
from typing import Optional

from src.config import settings
from src.utils.logger import logger

security = HTTPBearer()


async def verify_jwt_token(
    credentials: HTTPAuthorizationCredentials = None
) -> dict:
    """
    Verify JWT token from Authorization header
    
    Args:
        credentials: Bearer token credentials
        
    Returns:
        Decoded token payload
        
    Raises:
        HTTPException: If token is invalid or expired
    """
    if not credentials:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Authorization header missing",
            headers={"WWW-Authenticate": "Bearer"}
        )
    
    token = credentials.credentials
    
    try:
        # Decode JWT token
        payload = jwt.decode(
            token,
            settings.JWT_SECRET,
            algorithms=["HS256"]
        )
        
        # Verify required fields
        if "userId" not in payload:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Invalid token payload"
            )
        
        return payload
    
    except jwt.ExpiredSignatureError:
        logger.warning("Expired JWT token")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token has expired",
            headers={"WWW-Authenticate": "Bearer"}
        )
    
    except jwt.InvalidTokenError as e:
        logger.warning(f"Invalid JWT token: {e}")
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token",
            headers={"WWW-Authenticate": "Bearer"}
        )


def get_user_id_from_token(payload: dict) -> int:
    """Extract user ID from token payload"""
    return payload.get("userId")


def get_user_role_from_token(payload: dict) -> Optional[str]:
    """Extract user role from token payload"""
    return payload.get("role")
